Move loaders into a separate folder
