+++ src/modules/loader/__mocks__/loaders.ts
@@ -1,8 +1,6 @@
-import type { Context } from '../../..'
 import {
   unknownDocs,
   type FunctionDocumentation,
-  type ModuleFunctions,
   type VariableDocumentation
 } from '../../moduleTypes'
 
@@ -52,30 +50,3 @@ export const loadModuleBundleAsync = jest.fn((name: string) => {
 })
 
 export const loadModuleTabsAsync = jest.fn().mockResolvedValue([])
-
-async function initModuleContextAsync(moduleName: string, context: Context, loadTabs: boolean) {
-  // Load the module's tabs
-  if (!(moduleName in context.moduleContexts)) {
-    context.moduleContexts[moduleName] = {
-      state: null,
-      tabs: loadTabs ? await loadModuleTabsAsync(moduleName) : null
-    }
-  } else if (context.moduleContexts[moduleName].tabs === null && loadTabs) {
-    context.moduleContexts[moduleName].tabs = await loadModuleTabsAsync(moduleName)
-  }
-}
-
-export default async function loadSourceModules(
-  sourceModulesToImport: Set<string>,
-  context: Context,
-  loadTabs: boolean
-) {
-  const loadedModules = await Promise.all(
-    [...sourceModulesToImport].map(async moduleName => {
-      await initModuleContextAsync(moduleName, context, loadTabs)
-      const bundle = await loadModuleBundleAsync(moduleName)
-      return [moduleName, bundle] as [string, ModuleFunctions]
-    })
-  )
-  context.nativeStorage.loadedModules = Object.fromEntries(loadedModules)
-}
