+++ src/parser/__tests__/disallowed-syntax.ts
@@ -2,7 +2,7 @@ import { Chapter } from '../../types'
 import { stripIndent } from '../../utils/formatters'
 import { expectParsedError } from '../../utils/testing'
 
-jest.mock('../../modules/loader')
+jest.mock('../../modules/loader/loaders')
 
 test('Cannot leave blank init in for loop', () => {
   return expectParsedError(
